Add favorites manager as dependency on the views
We continue the goal to have favorites manager to be source of truth.
On this diff, the manager is created when initializing the app and its inject on all the views.
As we need to know the favorites before showing the breeds, we fetch the favorites before showing any view.

Signed-off-by: Rui Barbosa <rui.barbosa@hey.com>